distributed mode dirty fix, code to be cleaned

+++ src/main/java/com/emc/mongoose/base/data/impl/UniformDataSource.java
@@ -113,8 +113,8 @@ implements DataSource<T> {
 	////////////////////////////////////////////////////////////////////////////////////////////////
 	@Override
 	public final void writeExternal(final ObjectOutput out)
-		throws IOException {
-		out.writeLong(dataSrcDirect.array().length);
+	throws IOException {
+		out.writeInt(dataSrcDirect.array().length);
 		out.writeLong(seed);
 	}
 	//
